requirements.txt
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.1
torch==2.0.1
transformers==4.26.1
datasets==2.9.0
gradio==3.15.0
pytest==7.2.1
pyyaml==6.0
tqdm==4.64.1
nltk